+++ src/main/java/de/maxhenkel/audioplayer/interfaces/CustomSoundHolder.java
@@ -1,0 +1,9 @@
+package de.maxhenkel.audioplayer.interfaces;
+
+import de.maxhenkel.audioplayer.CustomSound;
+
+public interface CustomSoundHolder {
+
+    CustomSound soundplayer$getCustomSound();
+
+}